refactoring,   extract code into separate file
