+++ src/javascript/app/config.js
@@ -1,0 +1,9 @@
+const LocalStorageKeys = {
+	LOGIN_AND_SAVE_MESSAGE_SEEN: 'loginAndsaveMessageSeen',
+	ASKED_TO_IMPORT_CREATIONS: 'askedToImportCreations'
+};
+
+
+export { LocalStorageKeys };
+
+
